requirements.txt
langchain==0.3.7
langchain_community==0.3.7
langchain_core==0.3.20
langchain_openai==0.2.9
openai==1.55.0
python-dotenv==1.0.1
Requests==2.32.3
streamlit==1.38.0
transformers==4.39.3
